Move `TraceHandler` to internals of `org.eclipse.terminal.view.ui`
This type is not in fact used in other bundles therefore it should not
be a part of API
