+++ terminal/bundles/org.eclipse.terminal.view.ui/src/org/eclipse/terminal/view/ui/internal/TraceHandler.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2018 Wind River Systems, Inc. and others. All rights reserved.
+ * Copyright (c) 2011, 2025 Wind River Systems, Inc. and others. All rights reserved.
  * This program and the accompanying materials are made available under the terms
  * of the Eclipse Public License 2.0 which accompanies this distribution, and is
  * available at https://www.eclipse.org/legal/epl-2.0/
@@ -8,13 +8,13 @@
  *
  * Contributors:
  * Wind River Systems - initial API and implementation
+ * Alexander Fedorov (ArSysOp) - further evolution
  *******************************************************************************/
-package org.eclipse.terminal.view.core.utils;
+package org.eclipse.terminal.view.ui.internal;
 
 import org.eclipse.core.runtime.Assert;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Platform;
-import org.eclipse.terminal.view.core.internal.CoreBundleActivator;
 
 /**
  * Helper class to handle tracing using the platforms debug capabilities.
@@ -184,11 +184,11 @@ public class TraceHandler {
 	 * <p>
 	 * Initializes the tracing handler with the given bundle identifier.
 	 *
-	 * @param identifier The bundle identifier or <code>null</code>.
+	 * @param identifier The bundle identifier, must not be <code>null</code>.
 	 */
 	public TraceHandler(String identifier) {
-		this.identifier = identifier != null ? identifier : CoreBundleActivator.getUniqueIdentifier();
-		Assert.isNotNull(this.identifier);
+		Assert.isNotNull(identifier);
+		this.identifier = identifier;
 	}
 
 	/**
